Extract ChestActions out of MapActions
